refactor: 프로덕트 파일 패키지 구조화
