--- src/main/java/com/lgmpjt/websnsspringboot/user/UserPort.java
@@ -1,5 +1,0 @@
-package com.lgmpjt.websnsspringboot.user;
-
-interface UserPort {
-	void save(final User user);
-}
