--- src/main/java/com/lgmpjt/websnsspringboot/user/UserRepository.java
@@ -1,6 +1,0 @@
-package com.lgmpjt.websnsspringboot.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface UserRepository extends JpaRepository<User, Long> {
-}
