+++ src/main/java/com/lgmpjt/websnsspringboot/user/data/UserCreateDto.java
@@ -1,32 +1,14 @@
-package com.lgmpjt.websnsspringboot.user;
+package com.lgmpjt.websnsspringboot.user.data;
 
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 import org.springframework.util.Assert;
 
-@Entity
-@Table(name = "User")
-@Getter
-@Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class User {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long userSeq;
-
+public class UserCreateDto {
 	private String userId;
-
 	private String password;
-
 	private String userName;
-
 	private String userEmail;
 
-	public User(String userId, String password, String userName, String userEmail) {
+	public UserCreateDto(String userId, String password, String userName, String userEmail) {
 		Assert.hasText(userId, "유저 아이디 값은 필수입니다.");
 		Assert.notNull(password, "패스워드 값은 필수입니다.");
 		Assert.hasText(userName, "유저 이름 값은 필수입니다.");
@@ -34,5 +16,21 @@ public class User {
 		this.password = password;
 		this.userName = userName;
 		this.userEmail = userEmail;
+	}
+
+	public String getUserId() {
+		return userId;
+	}
+
+	public String getPassword() {
+		return password;
+	}
+
+	public String getUserName() {
+		return userName;
+	}
+
+	public String getUserEmail() {
+		return userEmail;
 	}
 }
