+++ src/main/java/com/lgmpjt/websnsspringboot/user/service/UserAdapter.java
@@ -1,5 +1,7 @@
-package com.lgmpjt.websnsspringboot.user;
+package com.lgmpjt.websnsspringboot.user.service;
 
+import com.lgmpjt.websnsspringboot.user.model.User;
+import com.lgmpjt.websnsspringboot.user.repository.UserRepository;
 import org.springframework.stereotype.Component;
 
 @Component
