+++ src/main/java/com/lgmpjt/websnsspringboot/user/service/UserPort.java
@@ -1,0 +1,7 @@
+package com.lgmpjt.websnsspringboot.user.service;
+
+import com.lgmpjt.websnsspringboot.user.model.User;
+
+interface UserPort {
+	void save(final User user);
+}
